% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_procresults.R
\docType{methods}
\name{dictionary,JD3_ProcResults-method}
\alias{dictionary,JD3_ProcResults-method}
\title{Title}
\usage{
\S4method{dictionary}{JD3_ProcResults}(object)

\S4method{dictionary}{JD3_ProcResults}(object)
}
\arguments{
\item{JD3_ProcResults}{}

\item{JD3_ProcResults}{}
}
\description{
Title

Title
}
