% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sa.R
\name{localPolynomialProperties}
\alias{localPolynomialProperties}
\title{Title}
\usage{
localPolynomialProperties(horizon, degree = 3, kernel = c("Henderson",
  "Uniform", "Biweight", "Triweight", "Tricube", "Gaussian", "Triangular",
  "Parabolic"), endpoints = c("DAF", "CC", "LC", "QL", "CQ", "CN"),
  ic = 4.5)
}
\arguments{
\item{ic}{}
}
\description{
Title
}
