% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sa.R
\docType{methods}
\name{saDecomposition,JD3_X11Plus-method}
\alias{saDecomposition,JD3_X11Plus-method}
\title{Title}
\usage{
\S4method{saDecomposition}{JD3_X11Plus}(object)
}
\arguments{
\item{JD3_X11Plus}{}
}
\description{
Title
}
