% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sa.R
\docType{methods}
\name{show,JD3_LP_Properties-method}
\alias{show,JD3_LP_Properties-method}
\title{Title}
\usage{
\S4method{show}{JD3_LP_Properties}(object)
}
\arguments{
\item{JD3_LP_Properties}{}
}
\description{
Title
}
