% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sa.R
\name{x11.cn}
\alias{x11.cn}
\title{Title}
\usage{
x11.cn(data, period, mul = T, trend.horizon = 6, trend.degree = 2,
  trend.kernel = c("Henderson", "BiWeight", "TriWeight", "TriCube",
  "Uniform", "Triangular", "Epanechnikov", "Trapezoidal"),
  seas.horizon = 3, seas.kernel = c("Trapezoidal", "Henderson",
  "BiWeight", "TriWeight", "TriCube", "Uniform", "Triangular",
  "Epanechnikov"), extreme.lsig = 1.5, extreme.usig = 2.5)
}
\arguments{
\item{extreme.usig}{}
}
\description{
Title
}
