% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sa.R
\name{x11.default}
\alias{x11.default}
\title{Title}
\usage{
x11.default(y, period, multiplicative = TRUE, trendLength = 13,
  trendDegree = 2, trendKernel = c("Henderson", "BiWeight",
  "TriWeight", "TriCube", "Uniform", "Triangular", "Epanechnikov",
  "Trapezoidal"), asymmetric = "CutAndNormalize", seas0 = c("S3X3",
  "S3X1", "S3X5", "S3X9", "S3X15"), seas1 = c("S3X3", "S3X1", "S3X5",
  "S3X9", "S3X15"), lsigma = 1.5, usigma = 2.5)
}
\arguments{
\item{usigma}{}
}
\description{
Title
}
