% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sa.R
\name{x11.lp}
\alias{x11.lp}
\title{Title}
\usage{
x11.lp(data, period, mul = true, trend.horizon = 6, trend.degree = 2,
  trend.kernel = c("Henderson", "BiWeight", "TriWeight", "TriCube",
  "Uniform", "Triangular", "Epanechnikov", "Trapezoidal"),
  trend.asymmetric.degree = 0, trend.asymmetric.params = 2/(sqrt(pi) *
  3.5), trend.asymmetric.timeliness = 0,
  trend.asymmetric.passBand = pi/8, seas.horizon = 3,
  seas.kernel = c("Trapezoidal", "Henderson", "BiWeight", "TriWeight",
  "TriCube", "Uniform", "Triangular", "Epanechnikov"),
  extreme.lsig = 1.5, extreme.usig = 2.5)
}
\arguments{
\item{extreme.usig}{}
}
\description{
Title
}
